Refactor codebase for more modular linkages implementation
This commit overhauls the entire codebase with a special focus on implementing the ASN-NAG linkage in a more modular way. It substitutes classes 'SailsAminoAcid' and 'SailsSugar' with respective structures and methods in the 'Sails' namespace, adding support for more flexible residue-residue linkages. Also, sails_find class now accepts prediction and working maps to enhance the identification of potential glycan attachment sites.