Enable download and unpacking of clipper-gemmi-wrapper in get_sources
The get_sources script has been updated to enable downloading and unpacking of clipper-gemmi-wrapper. Plus, the if statement was reformatted for proper functioning. Various changes were also made regarding the clipper/minimol target directories and I/O operations. Furthermore, the C++ standard was updated from 11 to 17, and the project name was changed from 'nanobind_example' to 'sails' in CMakeLists.txt.