Add Glycan and Glycosite structures
New Glycan structure is added in sails-glycan.h which represents glycan as an adjacency list. Functionality for basic graph operations like bfs and dfs are also included. The file also introduces a structure, Sugar, that acts as a node in the Glycan graph. Additionally, a Glycosite structure is defined in sails-model.h, indexing the model, chain, and residue.