Add glycan topology search functionality
This commit introduces a new method 'find_glycan_topology' that conducts a search for the topology of the glycan molecules within a given structure. The functionality relies on gemmi's neighbor search functionality to identify nearby atoms. Two new files 'sails-topology.cpp' and 'sails-topology.h' were added to implement this method.