Add linkage header and implementation files
This commit introduces the `sails-linkage.h` and `sails-linkage.cpp` files in the include and src directories respectively, housing the `Sails::Model` namespace. This namespace provides several methods for residue calculations, atom extraction, and monomer retrieval that pertain to the implementation of a linkage database. The `extend` method added applies linkage data to generate new glycan units and updates the structure's residue list accordingly.