Implement calculate_superposition method
A new method called calculate_superposition has been implemented in the Sails namespace of the sails-vector.cpp file. This method is meant to calculate the superposition of two sets of points represented as Vectors, throwing an exception if the sizes of these Vector sets don't match. Necessary headers for this implementation have been included in the sails-vector.h file.