Refactor and streamline code in sails-linkage.cpp
This commit changes the extend method in the Sails::Model class and the calculate_residue_translation function. The extend method has been updated to use a simplified syntax and make better use of the Gemmi library's capabilities. The calculate_residue_translation function was also modified to directly call the calculate_superposition function.