Add iterator methods to Glycan class
Two new iterator methods have been added to the Glycan class: 'begin' and 'end'. These methods return constant iterators to the beginning and end of the 'sugars' map respectively. They allow safe traversal of the map in a forward direction without modifying it. A slight modification to the ResidueDatabase was also done for internal consistency.