Refactor sails-utils module and enhance print_vector function
This commit includes a refactoring of sails-utils.cpp file, which includes improvements on the structure and the removal of unnecessary includes. Additionally, the print_vector function has been extended in order to accept a label parameter. This function now allows the user to print a vector with a label, thus providing more context. Minor modifications are also done in the header file sails-utils.h for consistency.