Update CMakeLists with new files and headers
A new C++ file, sails-refine.cpp, has been added to the CMakeLists.txt for building. Additionally, the global pattern for gathering the headers was updated to ${WRK_DIR}/src/include/*.h. The setting to expose these headers as public properties of the csails target was uncommented, allowing for proper installation.