Refactor Topology constructor and improve debug messaging
The Topology constructor of the Sails module has been refactored to include an initialisation of the neighbour search functionality. This makes the search functionality immediately available upon instance generation, thereby streamlining the code. Debug messaging has been improved in the 'find_residue_near_donor' and 'find_glycan_topology' methods for better troubleshooting. A 'set_structure' method was also added to allow re-setting of the structure and corresponding re-initialisation of the neighbour search.