Add progress bar and density feature to the model
The code now includes a progress bar to show the status in sequences processing. Also, the "density" feature in the model enables the program to recognize and react to density differences while extending the glycan. This improves the intelligence of the program, giving it more accurate and efficient processing of glycosites.