Update Glycan topology and remove erroneous sugars
In the Sails.cpp module, the Glycan topology has been updated after the proper extension to represent the current structure. Moreover, introduced a method to remove the erroneous sugars from the Glycan based on a threshold rscc score. This prevents any faulty sugars from being incorporated into the structure. The non-buildable terminal sugars are handled separately. The changes result in a more accurate and reliable structure for the Glycans.