Refactor Sails::Topology to use pointer for gemmi::Structure
The code has been updated to use pointer for gemmi::Structure in Sails::Topology, replacing the previous reference type. This was done across the methods initialise_neighbour_search and in the Topology constructor. Making changes to the way gemmi::Structure object is managed will allow for more control over its state and lifecycle throughout the code.