Change parameter types in Model methods
The parameter types for the `Model` methods in `sails-linkage.h` and `sails-linkage.cpp` have been updated. Specifically, `structure` parameter in `Model` constructor is now a pointer. Furthermore, `extend` method now takes a `Glycosite` instead of an `int` for `base_seqid`. Also, some variables in the `extend` function have been changed to map and vector to accommodate for this change.