Add find_previous_sugar method in sails-glycan.h
A new method, find_previous_sugar, has been added to the sails-glycan.h file. This method is designed to find the previous sugar molecule linked to a given sugar molecule within the adjacency list. It returns the corresponding donor sugar molecule if found, or std::nullopt if no previous sugar is found.