Add functionality to detect and remove erroneous sugars
A new function, remove_erroneous_sugars, is added to the sails.cpp code base. This function checks for sugars within a glycan that fall below a certain rscc_threshold indicating an error and subsequently removes them. It also updates the scoring mechanism in the program and introduces minor changes in file save methods.